Consolidated bit-for-bit math functions into a new physics_common lib.
