+++ components/scream/src/physics/p3/CMakeLists.txt
@@ -53,7 +53,7 @@ target_include_directories(p3 PUBLIC ${SCREAM_INCLUDE_DIRS})
 target_include_directories(p3 SYSTEM PUBLIC ${SCREAM_TPL_INCLUDE_DIRS} ${CIMEROOT}/src/share/include)
 set_target_properties(p3 PROPERTIES
   Fortran_MODULE_DIRECTORY ${SCREAM_F90_MODULES})
-target_link_libraries(p3 scream_share ${SCREAM_TPL_LIBRARIES})
+target_link_libraries(p3 physics_common scream_share ${SCREAM_TPL_LIBRARIES})
 
 if (NOT SCREAM_LIB_ONLY)
   add_subdirectory(tests)
